+++ src/components/ManageRoles.css
@@ -1,39 +1,230 @@
 .manage-roles-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
   padding: 2rem;
+}
+
+.manage-roles-header {
+  margin-bottom: 2rem;
+  text-align: center;
+  color: white;
 }
 
 .roles-list {
   display: flex;
   flex-direction: column;
   gap: 1rem;
-  max-width: 600px;
+  margin-bottom: 1rem;
+  width: 100%;
+  max-width: 800px;
 }
 
 .role-card {
   display: flex;
   justify-content: space-between;
   align-items: center;
-  padding: 1rem;
+  padding: 1rem 1.5rem;
   background: rgba(1, 1, 15, 0.575);
   backdrop-filter: blur(10px);
   border-radius: 8px;
   color: white;
+  width: 100%;
 }
 
-.user-name {
+.role-info {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.role-name {
   font-size: 1.1rem;
+  font-weight: 500;
 }
 
-.role-select {
+.role-description {
+  color: #aaa;
+  font-size: 0.9rem;
+}
+
+.role-permissions {
+  color: #30c645;
+  font-size: 0.9rem;
+}
+
+.role-actions {
+  display: flex;
+  gap: 0.5rem;
+}
+
+.edit-button {
+  background: none;
+  border: none;
+  color: #3498db;
+  cursor: pointer;
   padding: 0.5rem;
+  transition: color 0.2s ease;
+}
+
+.edit-button:hover {
+  color: #2980b9;
+}
+
+.delete-button {
+  background: none;
+  border: none;
+  color: #ff4444;
+  cursor: pointer;
+  padding: 0.5rem;
+  transition: color 0.2s ease;
+}
+
+.delete-button:hover {
+  color: #ff0000;
+}
+
+.add-role-button {
+  padding: 0.8rem;
+  background-color: #30c645;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 0.5rem;
+  width: 150px;
+  margin-top: 1rem;
+  transition: background-color 0.2s ease;
+}
+
+.add-role-button:hover {
+  background-color: #28a33a;
+}
+
+.role-form {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  width: 100%;
+  max-width: 800px;
+  background: rgba(1, 1, 15, 0.575);
+  backdrop-filter: blur(10px);
+  padding: 1.5rem;
+  border-radius: 8px;
+  margin-bottom: 1.5rem;
+}
+
+.form-group {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.form-group label {
+  color: white;
+  font-size: 0.9rem;
+}
+
+.form-group input,
+.form-group textarea,
+.form-group select {
+  padding: 0.8rem;
   border: none;
   border-radius: 4px;
   background-color: rgba(255, 255, 255, 0.1);
   color: white;
-  cursor: pointer;
 }
 
-.role-select option {
+.form-group select option {
   background-color: #1a1a1a;
+}
+
+.checkbox-group {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 1rem;
+}
+
+.checkbox-item {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
   color: white;
+}
+
+.form-buttons {
+  display: flex;
+  gap: 1rem;
+  margin-top: 1rem;
+}
+
+.submit-button {
+  padding: 0.8rem;
+  background-color: #30c645;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  flex: 1;
+  transition: background-color 0.2s ease;
+}
+
+.submit-button:hover {
+  background-color: #28a33a;
+}
+
+.cancel-button {
+  padding: 0.8rem;
+  background-color: #6c757d;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  flex: 1;
+  transition: background-color 0.2s ease;
+}
+
+.cancel-button:hover {
+  background-color: #5a6268;
+}
+
+.error-message {
+  color: #ff4444;
+  text-align: center;
+  margin-bottom: 1rem;
+  padding: 0.5rem;
+  background: rgba(255, 0, 0, 0.1);
+  border-radius: 4px;
+  width: 100%;
+  max-width: 800px;
+}
+
+.success-message {
+  color: #30c645;
+  text-align: center;
+  margin-bottom: 1rem;
+  padding: 0.5rem;
+  background: rgba(48, 198, 69, 0.1);
+  border-radius: 4px;
+  width: 100%;
+  max-width: 800px;
+}
+
+.loading {
+  text-align: center;
+  padding: 2rem;
+  color: #666;
+}
+
+.no-roles {
+  text-align: center;
+  padding: 2rem;
+  color: #666;
+  background: rgba(1, 1, 15, 0.3);
+  border-radius: 8px;
+  width: 100%;
+  max-width: 800px;
 }
