+++ src/components/UserList.css
@@ -1,4 +1,7 @@
 .user-list-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
   padding: 2rem;
 }
 
@@ -7,22 +10,26 @@
   flex-direction: column;
   gap: 1rem;
   margin-bottom: 1rem;
+  width: 100%;
+  max-width: 800px;
 }
 
 .user-card {
   display: flex;
   justify-content: space-between;
   align-items: center;
-  padding: 1rem;
+  padding: 1rem 1.5rem;
   background: rgba(1, 1, 15, 0.575);
   backdrop-filter: blur(10px);
   border-radius: 8px;
   color: white;
+  width: 100%;
 }
 
 .user-info {
   display: flex;
   flex-direction: column;
+  align-items: flex-start;
   gap: 0.5rem;
 }
 
@@ -46,6 +53,7 @@
   color: #ff4444;
   cursor: pointer;
   padding: 0.5rem;
+  transition: color 0.2s ease;
 }
 
 .delete-button:hover {
@@ -59,6 +67,8 @@
   padding: 0.5rem;
   background: rgba(255, 0, 0, 0.1);
   border-radius: 4px;
+  width: 100%;
+  max-width: 800px;
 }
 
 .loading {
@@ -73,6 +83,8 @@
   color: #666;
   background: rgba(1, 1, 15, 0.3);
   border-radius: 8px;
+  width: 100%;
+  max-width: 800px;
 }
 
 .refresh-button {
@@ -87,6 +99,7 @@
   justify-content: center;
   gap: 0.5rem;
   width: 150px;
+  transition: background-color 0.2s ease;
 }
 
 .refresh-button:hover {
@@ -96,4 +109,44 @@
 .refresh-button:disabled {
   background-color: #ccc;
   cursor: not-allowed;
+}
+
+
+/* Add styles for role selection dropdown */
+.role-select {
+  padding: 0.6rem;
+  border: none;
+  border-radius: 4px;
+  background-color: rgba(255, 255, 255, 0.1);
+  color: white;
+  cursor: pointer;
+  transition: background-color 0.2s ease;
+}
+
+.role-select:hover {
+  background-color: rgba(255, 255, 255, 0.2);
+}
+
+.role-select option {
+  background-color: #1a1a1a;
+}
+
+/* Add styles for success message */
+.success-message {
+  color: #30c645;
+  text-align: center;
+  margin-bottom: 1rem;
+  padding: 0.5rem;
+  background: rgba(48, 198, 69, 0.1);
+  border-radius: 4px;
+  width: 100%;
+  max-width: 800px;
+}
+
+/* Add these styles to your existing CSS file */
+
+.user-actions {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
 }
